fix sorting code for files in src/ according to its version number prefixed at the filename
This means we change the way we name the source files (.txt). This is due to
Linux doesn't store creation date, and we cannot just rely on modified date
as we might go back to edit very old post. So it will break things.

Yep, this is a safer way to have consistent result on sorted files listed on index.html